Extract Receipt proof config to a module && split const definitions from computation
