+++ ethereum_history_api/oracles/src/script/noir_fixtures/new_receipt_proof.ts
@@ -1,5 +1,5 @@
 import { Proof } from '../../ethereum/proof.js';
-import { ReceiptProofConfigM } from '../../noir/oracles/receiptOracle/encode.js';
+import { ReceiptProofConfigM } from '../../noir/oracles/common/proofConfig/receipt.js';
 import { createNewTopLevelProofInputFixture } from './new_proof.js';
 
 export function createNewReceiptProofFixture(proof: Proof): string {
