+++ ethereum_history_api/oracles/src/script/noir_fixtures/receipt_proof.ts
@@ -1,5 +1,5 @@
 import { Proof } from '../../ethereum/proof.js';
-import { LEGACY_MAX_RECEIPT_ENCODED_LEN, ReceiptProofConfigM } from '../../noir/oracles/receiptOracle/encode.js';
+import { LEGACY_MAX_RECEIPT_ENCODED_LEN, ReceiptProofConfigM } from '../../noir/oracles/common/proofConfig/receipt.js';
 import { BYTE_HEX_LEN } from '../../util/const.js';
 import { createProofFixture } from './proof.js';
